+++ ansibleplaybookgrapher/renderer/graphviz/__init__.py
@@ -1,10 +1,23 @@
+# Copyright (C) 2023 Mohamed El Mouctar HAIDARA
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
 import os
-from typing import Dict, List, Tuple, Set
+from typing import Dict, List, Set
 
 from ansible.utils.display import Display
 from graphviz import Digraph
 
-from ansibleplaybookgrapher import GraphVizPostProcessor
 from ansibleplaybookgrapher.graph import (
     PlaybookNode,
     PlayNode,
@@ -14,6 +27,7 @@ from ansibleplaybookgrapher.graph import (
     TaskNode,
 )
 from ansibleplaybookgrapher.renderer import Builder
+from ansibleplaybookgrapher.renderer.graphviz.postprocessor import GraphVizPostProcessor
 
 display = Display()
